Abuse-Elevation-Control-Mechanism-Setuid-Setgid-2
Sub-technique of:  T1548 An adversary may perform shell escapes or exploit vulnerabilities in an application with the setsuid or setgid bits to get code running in a different user’s context. On Linux or macOS, when the setuid or setgid bits are set for an application, the application will run with the privileges of the owning user or group respectively.